+++ airmise/export.py
@@ -19,24 +19,23 @@ from inspect import getfullargspec
 from types import FunctionType
 
 from lk_utils import fs
-from lk_utils.textwrap import join
-from lk_utils.textwrap import wrap
+from lk_utils import tw
 
 
 class T:
     Functions = t.Dict[str, t.Iterable[t.Union[FunctionType, t.Callable]]]
-    FallbackType = t.Literal["any", "str"]
+    FallbackType = t.Literal['any', 'str']
     PlainParamType = t.Literal[
-        "any",
-        "bool",
-        "dict",
-        "float",
-        "int",
-        "list",
-        "none",
-        "set",
-        "str",
-        "tuple",
+        'any',
+        'bool',
+        'dict',
+        'float',
+        'int',
+        'list',
+        'none',
+        'set',
+        'str',
+        'tuple',
     ]
 
 
@@ -58,20 +57,17 @@ def export_functions(funcs: T.Functions, output_path: str) -> None:
                 #   be careful if there are same function names, the last one -
                 #   takes effect.
     """
-    if output_path.endswith(".py"):
+    if output_path.endswith('.py'):
         _export_functions_to_file(funcs, output_path)
     else:
         raise NotImplementedError
 
 
-# -----------------------------------------------------------------------------
+# ------------------------------------------------------------------------------
 
 
-# noinspection PyUnresolvedReferences
-def _export_functions_to_file(
-    funcs: T.Functions, output_path: str
-) -> None:
-    classified_funcs = _classify_functions(funcs.values())  # noqa
+def _export_functions_to_file(funcs: T.Functions, output_path: str) -> None:
+    classified_funcs = _classify_functions(funcs.values())  # type: ignore
     # if len(classified_funcs) > 1:
     #     used_names = set()
     #     for xdict in classified_funcs.values():
@@ -82,92 +78,92 @@ def _export_functions_to_file(
     #                 )
     #             else:
     #                 used_names.add(func_name)
-    
+
     custom_funcnames = {id(v): k for k, v in funcs.items()}
     defined_funcnames = []
     show_module_name_divider_line = len(classified_funcs) > 1
-    
+
     out_rows = [
-        "import airmise as air",
-        "from functools import partial",
-        "from typing import Any  # noqa",
-        "",
-        "",
+        'import airmise as air',
+        'from functools import partial',
+        'from typing import Any  # noqa',
+        '',
+        '',
     ]
-    
+
     for key0 in sorted(classified_funcs.keys()):
         if show_module_name_divider_line:
             module_name = key0
             out_rows.append(
-                "# {} {}".format(
-                    "-" * (80 - 2 - len(module_name) - 1), module_name
+                '# {} {}'.format(
+                    '-' * (80 - 2 - len(module_name) - 1), module_name
                 )
             )
-            out_rows.append("")
-        
+            out_rows.append('')
+
         for key1 in sorted(classified_funcs[key0].keys()):
             func_info = classified_funcs[key0][key1]
-            func_name = custom_funcnames[func_info["id"]]
-            signature = func_info["signature"]
-            
+            func_name = custom_funcnames[func_info['id']]
+            signature = func_info['signature']
+
             out_rows.append(
-                "def {}({}) -> {}:  # noqa".format(
+                'def {}({}) -> {}:  # noqa'.format(
                     func_name,
-                    ", ".join(
+                    ', '.join(
                         filter(
                             None,
                             (
-                                ", ".join(
+                                ', '.join(
                                     (
                                         name
-                                        if name.startswith("*")
-                                        else "{}: {}".format(
+                                        if name.startswith('*')
+                                        else '{}: {}'.format(
                                             name, _translate_type(type_)
                                         )
                                     )
-                                    for name, type_ in signature["args"]
+                                    for name, type_ in signature['args']
                                 ),
-                                ", ".join(
+                                ', '.join(
                                     (
                                         name
-                                        if name.startswith("**")
-                                        else "{}: {} = {}".format(
+                                        if name.startswith('**')
+                                        else '{}: {} = {}'.format(
                                             name,
                                             _translate_type(type_),
                                             default,
                                         )
                                     )
                                     for name, type_, default in signature[
-                                        "kwargs"
+                                        'kwargs'
                                     ]
                                 ),
                             ),
                         )
                     ),
-                    _translate_type(signature["return"]),
+                    _translate_type(signature['return']),
                 )
             )
-            if func_info["document"]:
+            if func_info['document']:
                 out_rows.append(
                     '    """{}    """'.format(
-                        indent(func_info["document"], rstrip=False)
+                        tw.indent(func_info['document'], rstrip=False)
                     )
                 )
-            out_rows.append("    ...")
-            out_rows.append("")
-            out_rows.append("")
-            
+            out_rows.append('    ...')
+            out_rows.append('')
+            out_rows.append('')
+
             defined_funcnames.append(func_name)
-    
+
     out_rows.append(
-        wrap(
+        tw.wrap(
             """
             globals().update({{
                 {}
             }})
             """
         ).format(
-            join(
+            tw.join(
                 (
                     '"{}": partial(air.call, "{}"),'.format(x, x)
                     for x in sorted(set(defined_funcnames))
@@ -176,11 +172,11 @@ def _export_functions_to_file(
             )
         )
     )
-    
+
     fs.dump(out_rows, output_path)
 
 
-# -----------------------------------------------------------------------------
+# ------------------------------------------------------------------------------
 
 
 def _classify_functions(
@@ -194,7 +190,7 @@ def _classify_functions(
     out = defaultdict(dict)
     for f in funcs:
         info = _parse_function(f)
-        out[info["module"]][info["name"]] = info
+        out[info['module']][info['name']] = info
     return out
 
 
@@ -220,7 +216,7 @@ def _parse_function(func: FunctionType) -> dict:
         #   )
     """
     annotations = Annotations(spec.annotations)
-    
+
     args = []
     if spec.defaults:
         args_count = len(spec.args) - len(spec.defaults)
@@ -231,8 +227,8 @@ def _parse_function(func: FunctionType) -> dict:
         type_ = annotations.get_arg_type(name)
         args.append((name, type_))
     if spec.varargs:
-        args.append(("*" + spec.varargs, "any"))
-    
+        args.append(('*' + spec.varargs, 'any'))
+
     kwargs = []
     if spec.defaults:
         enum: t.Iterator[t.Tuple[int, int]] = enumerate(
@@ -244,32 +240,28 @@ def _parse_function(func: FunctionType) -> dict:
             type_ = annotations.get_kwarg_type(name, default)
             kwargs.append((name, type_, default))
     if spec.kwonlyargs:
-        for name, default in spec.kwonlydefaults.items():
+        for name, default in spec.kwonlydefaults.items():  # type: ignore
             type_ = annotations.get_kwarg_type(name, default)
             kwargs.append((name, type_, default))
     if spec.varkw:
-        kwargs.append(("**" + spec.varkw, "any", ...))
-    
+        kwargs.append(('**' + spec.varkw, 'any', ...))
+
     return_ = annotations.get_return_type()
-    
+
     return {
-        "id"       : id(func),
-        "module"   : func.__module__,
-        "name"     : func.__name__,
-        "document" : func.__doc__ or "",
-        "signature": {
-            "args"  : args,
-            "kwargs": kwargs,
-            "return": return_,
-        },
+        'id': id(func),
+        'module': func.__module__,
+        'name': func.__name__,
+        'document': func.__doc__ or '',
+        'signature': {'args': args, 'kwargs': kwargs, 'return': return_},
     }
 
 
 def _translate_type(type0: T.PlainParamType) -> str:
-    if type0 == "any":
-        return "Any"
-    elif type0 == "none":
-        return "None"
+    if type0 == 'any':
+        return 'Any'
+    elif type0 == 'none':
+        return 'None'
     else:
         return type0
 
@@ -278,80 +270,80 @@ class Annotations:
     def __init__(
         self,
         annotations: t.Dict[str, t.Any],
-        fallback_type: T.FallbackType = "any",
+        fallback_type: T.FallbackType = 'any',
     ) -> None:
         self.annotations = annotations
         self._fallback_type = fallback_type
         self._type_2_str = {
-            "any"    : "any",
-            "anystr" : "str",
-            "bool"   : "bool",
-            "dict"   : "dict",
-            "float"  : "float",
-            "int"    : "int",
-            "list"   : "list",
-            "literal": "str",
-            "none"   : "none",
-            "set"    : "set",
-            "str"    : "str",
-            "tuple"  : "tuple",
-            "union"  : "any",
+            'any': 'any',
+            'anystr': 'str',
+            'bool': 'bool',
+            'dict': 'dict',
+            'float': 'float',
+            'int': 'int',
+            'list': 'list',
+            'literal': 'str',
+            'none': 'none',
+            'set': 'set',
+            'str': 'str',
+            'tuple': 'tuple',
+            'union': 'any',
         }
         # if config.BARE_NONE_MEANS_ANY:
         #     self._type_2_str['none'] = 'any'
-    
+
     # noinspection PyUnresolvedReferences,PyProtectedMember
     def _normalize_type(self, type_: t.Any) -> T.PlainParamType:
         out = str(type_)
-        
+
         if isinstance(type_, str):
             pass
-        elif isinstance(type_, t._TypedDictMeta):
-            return "dict"
-        elif (x := getattr(type_, "__base__", None)) and str(
+        elif isinstance(type_, t._TypedDictMeta):  # type: ignore
+            return 'dict'
+        elif (x := getattr(type_, '__base__', None)) and str(
             x
         ) == "<class 'tuple'>":
-            return "tuple"  # typing.NamedTuple
+            return 'tuple'  # typing.NamedTuple
         else:
-            # if we are running in lower version python, be noted some classes -
+            # if we are running in lower version python, be noted some classes 
             # are not available.
             _is_legacy_typing: bool = (
-                getattr(t, "_LiteralGenericAlias", None) is None
+                getattr(t, '_LiteralGenericAlias', None) is None
             )
             if _is_legacy_typing:
-                if isinstance(type_, t._GenericAlias):
-                    return "any"
+                if isinstance(type_, t._GenericAlias):  # type: ignore
+                    return 'any'
             else:
-                if isinstance(type_, t._LiteralGenericAlias):
+                if isinstance(type_, t._LiteralGenericAlias):  # type: ignore
                     # e.g.
                     #   sometype = typing.Literal['A', 'B', 'C']
                     #   type(sometype)  # -> typing._LiteralGenericAlias
-                    return "str"
-                elif isinstance(type_, t._UnionGenericAlias):
+                    return 'str'
+                elif isinstance(type_, t._UnionGenericAlias):  # type: ignore
                     # e.g.
                     #   sometype = typing.Union[str, None]
                     #   type(sometype)  # -> typing._UnionGenericAlias
                     return self._normalize_type(type_.__args__[0])
-                elif isinstance(type_, t._GenericAlias):
+                elif isinstance(type_, t._GenericAlias):  # type: ignore
                     out = type_._name
-        
+
         assert isinstance(out, str)
         out = out.lower()
-        if out.startswith("<class "):
+        if out.startswith('<class '):
             out = out[8:-2]  # "<class 'list'>" -> "list"
-        if "[" in out:
-            out = out.split("[", 1)[0]
+        if '[' in out:
+            out = out.split('[', 1)[0]
         # print(':v', type_, out)
         if out in self._type_2_str:
-            return self._type_2_str[out]
-        return "any"
-    
+            return self._type_2_str[out]  # type: ignore
+        return 'any'
+
     def get_arg_type(self, name: str) -> T.PlainParamType:
         if name in self.annotations:
             return self._normalize_type(self.annotations[name])
         else:
             return self._fallback_type
-    
+
     def get_kwarg_type(self, name: str, value: t.Any) -> T.PlainParamType:
         if name in self.annotations:
             out = self._normalize_type(self.annotations[name])
@@ -359,23 +351,22 @@ class Annotations:
             out = self.deduce_type_by_value(value)
         # noinspection PyTypeChecker
         return out
-    
+
     def get_return_type(self) -> T.PlainParamType:
-        if "return" in self.annotations:
-            return self._normalize_type(self.annotations["return"])
+        if 'return' in self.annotations:
+            return self._normalize_type(self.annotations['return'])
         else:
-            return "none"
-    
+            return 'none'
+
     @staticmethod
     def deduce_type_by_value(default: t.Any) -> T.PlainParamType:
-        # noinspection PyTypeChecker
-        return {
-            bool : "bool",
-            dict : "dict",
-            float: "float",
-            int  : "int",
-            list : "list",
-            set  : "set",
-            str  : "str",
-            tuple: "tuple",
-        }.get(type(default), "any")
+        return {  # type: ignore
+            bool: 'bool',
+            dict: 'dict',
+            float: 'float',
+            int: 'int',
+            list: 'list',
+            set: 'set',
+            str: 'str',
+            tuple: 'tuple',
+        }.get(type(default), 'any')