+++ changelog.md
@@ -1,9 +1,13 @@
 # Airmise Changelog
 
+### ...
+
+- Show accurate position of error source.
+
 ### 3.1.0 (WIP)
 
 - Inject connection object into server holded functions.
-- Show accurate position of error source.
+- Proxy roles.
 
 ### 3.0.0 (2026-07-07)